+++ core/Cargo.toml
@@ -48,7 +48,6 @@ windows = { version = "0.44.0", features = [
 ] }
 
 [features]
-bad_host = []
 garbage_collect = []
 
 [lib]
